ExploratoryFactorAnalysis

Documentation for ExploratoryFactorAnalysis.

MultivariateStats.PCA

ExploratoryFactorAnalysis.EigenRefinement

ExploratoryFactorAnalysis.FactorAnalysisFull

ExploratoryFactorAnalysis.FactorAnalysisGN

ExploratoryFactorAnalysis.FactorAnalysisLAD

ExploratoryFactorAnalysis.FactorAnalysisMM

ExploratoryFactorAnalysis.FactorAnalysisPartial

ExploratoryFactorAnalysis.GenerateRandomData

ExploratoryFactorAnalysis.PositiveDefiniteProjection

ExploratoryFactorAnalysis.ProxAbsolute

ExploratoryFactorAnalysis.RunTests

MultivariateStats.PCA — Method

Forms r principal components of the data matrix X.

ExploratoryFactorAnalysis.EigenRefinement — Method

Refines r approximate eigenvectors of the symmetric matrix S.

ExploratoryFactorAnalysis.FactorAnalysisFull — Method

Performs rank r factor analysis on S by exact factor loading updates with a full eigen-decomposition.

ExploratoryFactorAnalysis.FactorAnalysisGN — Method

Performs rank r factor analysis on S by standard Gauss-Newton updates.

ExploratoryFactorAnalysis.FactorAnalysisLAD — Method

Performs rank r factor analysis on S by Gauss-Newton updates and an approximate least absolute deviation loss.

ExploratoryFactorAnalysis.FactorAnalysisMM — Method

Performs rank r factor analysis on S by MM Gauss-Newton updates.

ExploratoryFactorAnalysis.FactorAnalysisPartial — Method

Performs rank r factor analysis on S by exact factor loading updates with a partial eigen-decomposition.

ExploratoryFactorAnalysis.GenerateRandomData — Method

Generates random factor analysis data.

ExploratoryFactorAnalysis.PositiveDefiniteProjection — Method

Projects the symmetric matrix S onto the closest positive semidefinite matrix of rank r or less by full eigen-decomposition.

ExploratoryFactorAnalysis.ProxAbsolute — Method

Evaluates the proximal map (soft thresholding) of the absolute value |y|.

ExploratoryFactorAnalysis.RunTests — Method

Runs test problems of rank r and Moreau constant mu.